Fixed a lot of compiler warnings, most of them unused variables. The variables that are left need a better look before fixe up

+++ src/SoilHardware/eqep.h
@@ -68,7 +68,7 @@ public:
   void WaitForPositionChangeCancel() { this->threadRunning = false; }
 
   // Set the polling period
-  void set_period(uint64_t period);
+  void set_period(long long unsigned int period);
 
   // Get the polling period of the encoder
   uint64_t get_period();
